# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ServRedi FSM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /time-entries/{id}/clock-out:
    patch:
      operationId: clockOut
      tags: [timeTracking]
      summary: Clock out — end a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutBody"
      responses:
        "200":
          description: Ended time entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quotes
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Created quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a quote by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote with line items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteWithItems"
        "404":
          description: Not found
    patch:
      operationId: updateQuote
      tags: [quotes]
      summary: Update a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteBody"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotes/{id}/items:
    post:
      operationId: addQuoteItem
      tags: [quotes]
      summary: Add a line item to a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteItemBody"
      responses:
        "201":
          description: Created quote item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteItem"
  /quotes/{id}/items/{itemId}:
    patch:
      operationId: updateQuoteItem
      tags: [quotes]
      summary: Update a quote line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteItemBody"
      responses:
        "200":
          description: Updated quote item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteItem"
    delete:
      operationId: deleteQuoteItem
      tags: [quotes]
      summary: Delete a quote line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
  /reports/employee-hours:
    get:
      operationId: getEmployeeHoursReport
      tags: [reports]
      summary: Employee hours report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: technicianId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Employee hours data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeHoursRow"
  /reports/jobs-summary:
    get:
      operationId: getJobsSummaryReport
      tags: [reports]
      summary: Jobs summary report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Jobs summary data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsSummaryReport"
...
components:
  schemas:
...
    TimeEntry:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        technicianId:
          type: integer
        technicianName:
          type: string
        jobId:
          type: ["integer", "null"]
        jobTitle:
          type: ["string", "null"]
        taskId:
          type: ["integer", "null"]
        clockInAt:
          type: string
        clockOutAt:
          type: ["string", "null"]
        breakMinutes:
          type: integer
        totalMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - organizationId
        - technicianId
        - technicianName
        - clockInAt
        - breakMinutes
        - isActive
        - createdAt
        - updatedAt
...
    ClockOutBody:
      type: object
      properties:
        notes:
          type: string
        latitude:
          type: number
        longitude:
          type: number
...
    Quote:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        notes:
          type: ["string", "null"]
        taxRate:
          type: number
        subtotal:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - organizationId
        - customerId
        - customerName
        - status
        - taxRate
        - subtotal
        - taxAmount
        - total
        - createdAt
        - updatedAt
    QuoteWithItems:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        notes:
          type: ["string", "null"]
        taxRate:
          type: number
        subtotal:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - organizationId
        - customerId
        - customerName
        - status
        - taxRate
        - subtotal
        - taxAmount
        - total
        - items
        - createdAt
        - updatedAt
    CreateQuoteBody:
      type: object
      properties:
        customerId:
          type: integer
        notes:
          type: string
        taxRate:
          type: number
      required:
        - customerId
    UpdateQuoteBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, approved, rejected]
        notes:
          type: string
        taxRate:
          type: number
    QuoteItem:
      type: object
      properties:
        id:
          type: integer
        quoteId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        quantity:
          type: number
        unitCost:
          type: number
        laborHours:
          type: ["number", "null"]
        laborRate:
          type: ["number", "null"]
        materialCost:
          type: ["number", "null"]
        lineTotal:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - quoteId
        - name
        - quantity
        - unitCost
        - lineTotal
        - createdAt
        - updatedAt
    CreateQuoteItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitCost:
          type: number
        laborHours:
          type: number
        laborRate:
          type: number
        materialCost:
          type: number
      required:
        - name
        - quantity
        - unitCost
    UpdateQuoteItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitCost:
          type: number
        laborHours:
          type: number
        laborRate:
          type: number
        materialCost:
          type: number
...
    EmployeeHoursRow:
      type: object
      properties:
        technicianId:
          type: integer
        technicianName:
          type: string
        totalHours:
          type: number
        breakHours:
          type: number
        netHours:
          type: number
        jobsWorked:
          type: integer
        entries:
          type: integer
      required:
        - technicianId
        - technicianName
        - totalHours
        - breakHours
        - netHours
        - jobsWorked
        - entries
    JobsSummaryReport:
      type: object
      properties:
        totalJobs:
          type: integer
        completedJobs:
          type: integer
        cancelledJobs:
          type: integer
        inProgressJobs:
          type: integer
        scheduledJobs:
          type: integer
        completionRate:
          type: number
        statusBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        priorityBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/PriorityCount"
      required:
        - totalJobs
        - completedJobs
        - cancelledJobs
        - inProgressJobs
        - scheduledJobs
        - completionRate
        - statusBreakdown
        - priorityBreakdown
    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count
    PriorityCount:
      type: object
      properties:
        priority:
          type: string
        count:
          type: integer
      required:
        - priority
        - count